Add files via upload
These files relate to my Capstone Project in Data Science. It was about finding the patterns in the texts and blogs of popular travel bloggers who help the Travel Industry to flourish. I took the example case of "meetsouthafrica" campaign that was a very successful social media project carried out in collaboration between 15 travel bloggers and tSouth Africa tourism. I performed text mining including topic modelling to find out the general topics that the travel bloggers discuss about in their blogs or tweet, sentiment analysis to find the sentiments that they usually maintain in the blogs and tweets, image analysis to find whether they post real pictures or photoshopped ones in order to grab the attention of the readers and network analysis to find the communities among the popular travel bloggers. The network analysis has been done in Gephi and NodeXL Pro and hence the code files have not been added here.